requirements.txt
numpy
streamlit
psycopg2-binary
APScheduler
